pyproject.toml
[tool.poetry]
name = "qr-code-generator"
version = "0.1.0"
description = "A simple QR code generator web application."
authors = ["Balazs Vagvolgyi <vagvolgyibalazsrobert@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
qrcode = "^7.3.1"
Pillow = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
